refactor: adjust hoc
