+++ src/widgets/antd/date.jsx
@@ -8,8 +8,9 @@ import { DatePicker, TimePicker } from 'antd';
 import dateHoc from '../../components/dateHoc';
 
 export default function date(p) {
+  const { format = 'dateTime' } = p.schema;
   const onChange = (value, string) => p.onChange(p.name, string);
-  const DateComponent = p.format === 'time' ? TimePicker : DatePicker;
+  const DateComponent = format === 'time' ? TimePicker : DatePicker;
   const FormDate = dateHoc(p, onChange)(DateComponent);
   return <FormDate />;
 }
