+++ src/widgets/antd/multiSelect.jsx
@@ -4,9 +4,10 @@
  */
 
 import React from 'react';
-import rangeHoc from '../../components/rangeHoc';
+import { Select } from 'antd';
+import multiSelectHoc from '../../components/multiSelectHoc';
 
 export default function multiSelect(p) {
-  const FormMulti = rangeHoc(p)(Select);
+  const FormMulti = multiSelectHoc(p)(Select);
   return <FormMulti />;
 }
