+++ src/widgets/antd/select.jsx
@@ -1,3 +1,8 @@
+/**
+ * Updated by Tw93 on 2019-12-07.
+ * 选择组件
+ */
+
 import React from 'react';
 import { Select } from 'antd';
