+++ src/widgets/fusion/date.jsx
@@ -5,12 +5,12 @@ import dateHoc from '../../components/dateHoc';
 import { getFormat } from '../../base/utils';
 
 export default function date(p) {
-  const { format } = p.schema;
+  const { format = 'dateTime' } = p.schema;
   const dateFormat = getFormat(format);
   const onChange = value => {
     p.onChange(p.name, moment(value || '', dateFormat).format(dateFormat));
   };
-  const DateComponent = p.format === 'time' ? TimePicker : DatePicker;
-  const FormDate = dateHoc({ p, onChange })(DateComponent);
+  const DateComponent = format === 'time' ? TimePicker : DatePicker;
+  const FormDate = dateHoc(p, onChange)(DateComponent);
   return <FormDate />;
 }
